183: Changing folder structure Moved Layout to hoc and created a folder for Auxiliary
