+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Aux from '../../../hoc/Auxiliary';
+import Aux from '../../../hoc/Aux/Auxiliary';
 import Button from '../../UI/Button/Button';
