+++ src/components/Navigation/SideDrawer/SideDrawer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styleClasses from './SideDrawer.module.css';
 import Logo from '../../Logo/Logo';
 import NavigationItems from '../NavigationItems/NavigationItems';
-import Aux from '../../../hoc/Auxiliary';
+import Aux from '../../../hoc/Aux/Auxiliary';
 import Backdrop from '../../UI/Backdrop/Backdrop';
 
 const SideDrawer = (props) => {
