+++ src/components/UI/Modal/Modal.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import styleClasses from './Modal.module.css';
-import Aux from '../../../hoc/Auxiliary';
+import Aux from '../../../hoc/Aux/Auxiliary';
 import Backdrop from '../Backdrop/Backdrop';
 
 class Modal extends Component {
