+++ src/containers/BurgerBuilder/BurgerBuilder.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Burger from '../../components/Burger/Burger';
 import Modal from "../../components/UI/Modal/Modal";
-import Aux from '../../hoc/Auxiliary';
+import Aux from '../../hoc/Aux/Auxiliary';
 import BuildControls from '../../components/Burger/BuildControls/BuildControls';
 import OrderSummary from '../../components/Burger/OrderSummary/OrderSummary';
