+++ src/hoc/Layout/Layout.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import Aux from '../../hoc/Auxiliary';
-import Toolbar from '../Navigation/Toolbar/Toolbar';
+import Aux from '../Aux/Auxiliary';
+import Toolbar from '../../components/Navigation/Toolbar/Toolbar';
 import styleClasses from './Layout.module.css';
-import SideDrawer from '../Navigation/SideDrawer/SideDrawer';
+import SideDrawer from '../../components/Navigation/SideDrawer/SideDrawer';
 
 class Layout extends Component {
     state = {
